+++ ft_lstsize_bonus.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_lstsize_bonus.c                                 :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: josefelghnam <josefelghnam@student.42.fr>  +#+  +:+       +#+        */
+/*   By: jel-ghna <jel-ghna@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2025/04/19 20:36:21 by josefelghnam      #+#    #+#             */
-/*   Updated: 2025/04/19 20:47:15 by josefelghnam     ###   ########.fr       */
+/*   Created: 2025/04/19 20:36:21 by josefelghna       #+#    #+#             */
+/*   Updated: 2025/04/23 18:53:54 by jel-ghna         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,10 +14,10 @@
 
 int	ft_lstsize(t_list *lst)
 {
-	long	size;
+	int	size;
 
 	size = 0;
 	while (lst && ++size)
 		lst = lst->next;
-	return ((int)size);
+	return (size);
 }
